[feat] full ES6+ modules
What?: ref PR #13

Breaking Changes?
