+++ src/pages/home/main.js
@@ -1,8 +1,4 @@
-// eslint-disable-next-line import/no-unresolved
 import 'normalize.css/normalize.css';
 import '../../css/main.css';
 import './page.css';
 import '../../js/View/home';
-import Card from '../../js/View/Card';
-
-window.customElements.define('my-card', Card);
